feat(mulmoScript): phase-3a — ops layer moves into @mulmoclaude/mulmoscript-plugin/server
The entire mulmoScript ops layer — probes, renders, uploads, the
movie/PDF pipelines, the edge-triggered generation tracker, the dispatch
router, mulmoErrorCapture, and the realpath containment guards — moves
from the host into a new Node-only `./server` entry of the shared
package (mulmocast + graphai become peers), so phase 3b can wire
MulmoTerminal against the SAME implementation instead of duplicating
~700 lines of orchestration.

Host-specific transport is injected via MulmoScriptServerBackend:
stories dir, artifacts FileOps, hardened atomic write, the ffmpeg
probe, the logger, and the onGenerationEvent fan-out. MulmoClaude's
binding is the new server/plugins/mulmoscript-server.ts — it creates
the ops instance, registers the built-in dispatch through the package's
createMulmoScriptDispatchHandler, and wires the pubsub publisher; the
REST routes consume the same instance. Deleted from the host:
mulmo-script-ops.ts, mulmoscript-builtin.ts,
events/mulmoscript-generation.ts, utils/mulmoErrorCapture.ts (their
tests moved into the package, plus new edge-trigger tracker coverage).

Behavior-preserving relocation; the only intended delta is the log
prefix for captured GraphAI provider errors ("mulmocast" →
"mulmo-script" via the host log adapter). Package bumps to 0.2.0
(new subpath entry + peers); launcher dep ratchets to ^0.2.0.
