+++ packages/mulmoclaude/package.json
@@ -42,7 +42,7 @@
     "@mulmoclaude/google-plugin": "^0.2.1",
     "@mulmoclaude/html-plugin": "^0.2.5",
     "@mulmoclaude/markdown-plugin": "^0.1.10",
-    "@mulmoclaude/mulmoscript-plugin": "^0.1.0",
+    "@mulmoclaude/mulmoscript-plugin": "^0.2.0",
     "@mulmoclaude/spotify-plugin": "^0.1.1",
     "@mulmoclaude/x-plugin": "^0.1.2",
     "@receptron/task-scheduler": "^0.1.0",
