+++ packages/plugins/mulmoscript-plugin/vite.config.ts
@@ -3,13 +3,16 @@ import vue from "@vitejs/plugin-vue";
 import tailwindcss from "@tailwindcss/vite";
 import { resolve } from "node:path";
 
-// Two entries: the server-facing `.` core (save/update executes + tool
-// definition, imported by the host server build) and the browser `./vue`
-// (View/Preview). `vue` + `gui-chat-protocol/vue` are externalised so the
-// plugin and host share ONE instance (the injected PLUGIN_RUNTIME_KEY Symbol
-// must match); `@mulmocast/*` are externalised so both hosts resolve their
-// own copies (version lockstep with `mulmocast`). Mirrors html-plugin's
-// config. Declarations are emitted by vue-tsc (tsconfig.build.json) because
+// Three entries: the browser-safe `.` core (save/update executes + tool
+// definition), the browser `./vue` (View/Preview), and the Node-only
+// `./server` (mulmocast ops + dispatch router). `vue` +
+// `gui-chat-protocol/vue` are externalised so the plugin and host share ONE
+// instance (the injected PLUGIN_RUNTIME_KEY Symbol must match);
+// `@mulmocast/*` / `mulmocast` / `graphai` are externalised so both hosts
+// resolve their own single hoisted copies (GraphAILogger state is
+// module-local — a bundled second copy would silently break the error
+// capture). Node built-ins are externalised for the server entry.
+// Declarations are emitted by vue-tsc (tsconfig.build.json) because
 // vite-plugin-dts and the Vue SFC transform disagree about .vue d.ts paths.
 export default defineConfig({
   plugins: [vue(), tailwindcss()],
@@ -18,13 +21,25 @@ export default defineConfig({
       entry: {
         index: resolve(__dirname, "src/index.ts"),
         vue: resolve(__dirname, "src/vue/index.ts"),
+        server: resolve(__dirname, "src/server/index.ts"),
       },
       name: "GUIChatPluginMulmoScript",
       formats: ["es", "cjs"],
       fileName: (format, entryName) => `${entryName}.${format === "es" ? "js" : "cjs"}`,
     },
     rollupOptions: {
-      external: ["vue", "gui-chat-protocol", "gui-chat-protocol/vue", "@mulmocast/types", "@mulmocast/deck-web"],
+      external: [
+        "vue",
+        "gui-chat-protocol",
+        "gui-chat-protocol/vue",
+        "@mulmocast/types",
+        "@mulmocast/deck-web",
+        "mulmocast",
+        "graphai",
+        "fs",
+        "path",
+        "node:async_hooks",
+      ],
       output: {
         exports: "named",
         globals: { vue: "Vue" },
